+++ tests/__init__.py
@@ -1,0 +1,10 @@
+# -*- coding: utf-8 -*-
+
+import os
+
+
+# Directory where the example docs are located for testing
+TEST_DOCS_SRC = os.path.join(
+    os.path.dirname(os.path.abspath(__file__)),
+    'example'
+)
